Consistent search pages
Made pages with search consistent,
and also added search underneath the title
