+++ frontend/src/components/filters/role-filter.component.module.scss
@@ -1,41 +1,21 @@
 @use '@/styles/globals';
 
-.radioGroup {
-  display: flex;
-}
-
 .radioDropdownButton {
-  display: flex;
-  align-items: center;
   padding: 0;
-  background: none;
-  border: none;
   box-shadow: none;
 }
 
 .dropdownButton {
-  display: flex;
-  align-items: center;
-  padding: 0 globals.$spacing-s;
   height: 100%;
-  box-shadow: none;
-  color: inherit;
-  background: none;
-  border: none;
 }
 
 .dropdownButtonSelected {
-  color: #fff;
+  color: globals.$color-white;
   border-color: globals.$primary-color;
 }
 
-.selectedRoleLabel {
-  font-weight: 500;
-}
-
 .dropdownSelected {
-  color: #fff;
-  font-weight: 600;
+  color: globals.$color-white;
 }
 
 .dropdownSelectedBg {
