--- frontend/src/pages/data-products/data-products.module.scss
@@ -1,4 +1,0 @@
-.container {
-    height: 100%;
-    width: 100%;
-}
