+++ frontend/src/pages/people/people-table.module.scss
@@ -1,0 +1,5 @@
+@use '@/styles/globals';
+
+.row {
+  cursor: pointer;
+}
