+++ frontend/src/routes.tsx
@@ -9,13 +9,14 @@ import Cart from '@/pages/cart/cart.page.tsx';
 import { DataProduct } from '@/pages/data-product/data-product.page.tsx';
 import { DataProductCreate } from '@/pages/data-product-create/data-product-create.page.tsx';
 import { DataProductEdit } from '@/pages/data-product-edit/data-product-edit.page.tsx';
-import { DataProducts } from '@/pages/data-products/data-products.page.tsx';
+import { DataProductsTable } from '@/pages/data-products/data-products-table.component.tsx';
 import { Dataset } from '@/pages/dataset/dataset.page.tsx';
 import { DatasetEdit } from '@/pages/dataset-edit/dataset-edit.page.tsx';
 import { ErrorRootElement } from '@/pages/error/error-root-element.page.tsx';
 import { ExplorerPage } from '@/pages/explorer/explorer.page.tsx';
 import { Home } from '@/pages/home/home.page.tsx';
 import { Marketplace } from '@/pages/marketplace/marketplace.page.tsx';
+import { PeoplePage } from '@/pages/people/people-table.component.tsx';
 import { ApplicationPaths } from '@/types/navigation';
 import ProtectedRoute from './components/layout/protected/protected.layout.tsx';
 import { DataOutput } from './pages/data-output/data-output.page.tsx';
@@ -25,7 +26,6 @@ import { EnvironmentConfigCreate } from './pages/environment-config-create/envir
 import { EnvironmentConfigs } from './pages/environment-configs/environment-configs.page.tsx';
 import { EnvironmentCreate } from './pages/environment-create/environment-create.page.tsx';
 import { Environments } from './pages/environments/environments.page.tsx';
-import { People } from './pages/people/people.page.tsx';
 import { PlatformServiceConfig } from './pages/platform-service-config/platform-service-config.page.tsx';
 import { PlatformServiceConfigCreate } from './pages/platform-service-config-create/platform-service-config-create.page.tsx';
 import { PlatformsConfigs } from './pages/platforms-configs/platforms-configs.page.tsx';
@@ -54,7 +54,7 @@ const router = createBrowserRouter([
                         path: ApplicationPaths.DataProducts,
                         children: [
                             {
-                                element: <DataProducts />,
+                                element: <DataProductsTable />,
                                 index: true,
                             },
                             {
@@ -107,7 +107,7 @@ const router = createBrowserRouter([
                     },
                     {
                         path: ApplicationPaths.People,
-                        element: <People />,
+                        element: <PeoplePage />,
                     },
                     {
                         path: ApplicationPaths.AuditLogs,
